Reorganize insertion and opti some code. Start rework of tests.

+++ crates/services/txpool_v2/src/storage/mod.rs
@@ -87,5 +87,5 @@ pub trait Storage {
     ) -> Result<StorageData, Error>;
 
     /// Count the number of transactions in the storage.
-    fn count(&self) -> u64;
+    fn count(&self) -> usize;
 }
